% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmagmaR-package.R
\docType{package}
\name{hmagmaR-package}
\alias{hmagmaR}
\alias{hmagmaR-package}
\title{hmagmaR: Incorporation of chromatin interaction and epigenetic profiles to predict risk genes}
\description{
Combines chromatin interaction (e.g. PLAC-seq/Hi-C) data with regulatory region annotations (e.g. promoters, enhancers, H3K27ac) to link non-coding GWAS SNPs to their putative target genes, producing MAGMA-compatible gene annotation files for downstream gene-level association testing.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/aydanasg/hmagmaR}
  \item Report bugs at \url{https://github.com/aydanasg/hmagmaR/issues}
}

}
\author{
\strong{Maintainer}: Aydan Asgarova \email{aydan.asgarova00@gmail.com}

}
\keyword{internal}
