% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampledDownAnnotation.R
\name{SampledDownAnnotation}
\alias{SampledDownAnnotation}
\title{Gene level analysis carried out by magma (https://journals.plos.org/ploscompbiol/article?id=10.1371/journal.pcbi.1004219)}
\usage{
SampledDownAnnotation(
  hic,
  regulatoryRegions,
  snps,
  annotated_genes,
  snpgeneexon,
  loopNumber,
  sampleDownNumber,
  AnnotationFile
)
}
\arguments{
\item{hic}{Dataframe of chromatin interaction data with 6 columns in the format: "chrom1", "start1", "end1", "chrom2", "start2", "end2". Chromosome columns should have "chr" before the number}

\item{regulatoryRegions}{Dataframe of regulatory regions (H3K27ac) in the .bed format (chr, start, end). Chromosome columns should have "chr" before the number}

\item{snps}{Dataframe of reference snps. This could be the ".bim" file from g1000 reference genome in the format: "chr", "position", "rsid"}

\item{annotated_genes}{Dataframe of reference genes. This should contain columns with names "chr", "start", "end",    "ensg". Chromosome columns should have "chr" before the number}

\item{snpgeneexon}{Dataframe of genes and snps within it. This should contain columns with names "rsid", "ensg"}

\item{loopNumber}{Number to sample down the loop number to. This is useful when comparing more than one sample to ensure that the difference is not down to the difference in loop number}

\item{sampleDownNumber}{Number of times sample down the loop number randomly}

\item{AnnotationFile}{Path where to save the sampled down annotation files}
}
\value{
Function generates sampled down gene level analysis files (.genes.raw, .genes.out, .log.suppl, .log)
}
\description{
Generates an annotation file for magma input
}
\examples{
data("test_dataset.microglia.hg19")
out <- tempfile()
SampledDownAnnotation(
  hic = hic, regulatoryRegions = regulatoryRegions, snps = snps,
  annotated_genes = annotated_genes, snpgeneexon = snpgeneexon,
  loopNumber = 5000, sampleDownNumber = 1, AnnotationFile = out
)
}
