% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationFileHmagma.R
\name{AnnotationFileHmagma}
\alias{AnnotationFileHmagma}
\title{Gene level analysis carried out by magma (https://journals.plos.org/ploscompbiol/article?id=10.1371/journal.pcbi.1004219)}
\usage{
AnnotationFileHmagma(
  fileName,
  hic,
  regulatoryRegions,
  snps,
  AnnotationFile.path
)
}
\arguments{
\item{fileName}{Name in the ouput file}

\item{hic}{Chromatin interaction data file with 6 columns in the format: "chrom1", "start1", "end1", "chrom2", "start2", "end2". Chromosome columns should have "chr" before the number}

\item{regulatoryRegions}{File of regulatory regions (H3K27ac) in the .bed format (chr, start, end). Chromosome columns should have "chr" before the number}

\item{snps}{File of reference snps. This could be the ".bim" file from g1000 reference genome in the format: "chr", "position", "rsid"}

\item{AnnotationFile.path}{Path where to save the sampled down annotation files}
}
\value{
Function generates gene level analysis files (.genes.raw, .genes.out, .log.suppl, .log)
}
\description{
Generates an annotation file for magma input
}
