% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneLevelAnalysis_hmagma.R
\name{GeneLevelAnalysis_hmagma}
\alias{GeneLevelAnalysis_hmagma}
\title{Gene level analysis carried out by magma (https://journals.plos.org/ploscompbiol/article?id=10.1371/journal.pcbi.1004219)}
\usage{
GeneLevelAnalysis_hmagma(g1000, gwas, AnnotationFile, output)
}
\arguments{
\item{g1000}{Path to 1,000 Genomes reference data files (.bin, .bed, .fam, .synonyms). Include the name of the of the files in the path (e.g. /path_to_g1000/g1000_eur). Obtain from https://cncr.nl/research/magma/ under Auxiliary files.}

\item{gwas}{Path to the GWAS summary statistics to be used where rsid column name is "SNP", p-value column name is "P", and sample size column name is "N"}

\item{AnnotationFile}{Annotation file generated using AnnotationFileHmagma() or SampledDownAnnotation()}

\item{output}{Name of the output file}
}
\value{
Function generates gene level analysis files (.genes.raw, .genes.out, .log.suppl, .log)
}
\description{
Generates an annotation file for magma input
}
